+++ src/components/ids-masthead/demos/index.ts
@@ -1,2 +1,3 @@
 // Supporting components
 import '../ids-masthead';
+import '../../ids-layout-flex/ids-layout-flex';